requirements.txt
customtkinter
pandas
tkcalendar
tkinter
hashlib
tkinter.messagebox
datetime
mysql-connector-python
